+++ Percursus.lua
@@ -80,7 +80,6 @@ end
 function PercursusFrame:ADDON_LOADED(_, addOnName)
 	if addOnName == addonName then
 		local dbInit = Utils:InitializeDatabase()
-		Utils:InitializeUpdateNotice()
 		Utils:InitializeMinimapButton()
 		Options:Initialize()
 		RaceTracker:Initialize()