+++ core/Options.lua
@@ -47,6 +47,7 @@ local function GetVal(setting) return setting:GetValue() end
 
 function Options:Initialize()
 	local category, layout = Settings.RegisterVerticalLayoutCategory(addonName)
+	Addon:SetChangelog(PER.CHANGELOG_TEXT)
 
 	layout:AddInitializer(CreateSettingsListSectionHeaderInitializer(L["options.general"]))
 
@@ -231,6 +232,7 @@ function Options:Initialize()
 
 	-- About Section
 	AWL.Settings:AddAboutSection(layout, addonName)
+	Addon:AddChangelogButton(layout)
 
 	Settings.RegisterAddOnCategory(category)