+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Added: Race times for the event 'Northrend Cup'
-- Removed: TOC version for patch 12.0.7 [retail]
+- Added: Changelog window available from the options menu
+- Removed: Version notice chat messages
+- Adapted to the latest version of Arcane Wizard: Library to ensure full compatibility